Fix warnings for lnx32e

+++ algorithms/kernel/svm/svm_train_result.h
@@ -66,7 +66,7 @@ public:
         {
             if (_alpha[i] > zero) nSV++;
         }
-        printf(">> nSV %lu\n", nSV);
+        printf(">> nSV %d\n", (int)nSV);
 
         model.setNFeatures(xTable.getNumberOfColumns());
         services::Status s;
